Vitest batch 30: 3-port multi-batch — tour + chat-threads + wearables-bp-merge
First multi-port batch (switching to bundled batches for throughput —
single-file PRs were too slow for the ~25 remaining puppeteer tests).

Three ports, each split Vitest + slim puppeteer DOM file:

  - test-tour.js (70 asserts) → Vitest: tour.js structure, TOUR_STEPS
    content, target-not-found logic, viewport-clamping math, CSS, and
    main/views/settings/SW wiring. test-tour-dom.js (85 asserts) keeps
    the live overlay/spotlight/tooltip creation, step navigation,
    z-index computed styles, and full walkthrough.

  - test-chat-threads.js (73 asserts) → Vitest: window exports, state
    shape, thread CRUD, legacy migration, save/load round-trip, 50-thread
    pruning, backup snapshot, encryption patterns, ensureActiveThread,
    personality inheritance. test-chat-threads-dom.js (17 asserts) keeps
    HTML structure + getComputedStyle, rail-toggle classList, search-filter
    rendered-item readback.

  - test-wearables-bp-merge.js (15 asserts) → Vitest: pure source-inspection
    regexes. test-wearables-bp-merge-dom.js (3 asserts) keeps the live
    openManualLogForm idempotency probe.

Three env-artifact fixes in test-chat-threads (all noted inline):
  1. generateThreadId() is `'t_'+Date.now().toString(36)` — no counter.
     Rapid createNewThread() calls in Node (no DOM-render delay) collide
     on id. Added a 2ms gap before each call. The latent collision in
     generateThreadId itself is a real bug worth a follow-up.
  2. buildBackupSnapshot() early-returns null without `labcharts-profiles`
     — seed a minimal profile registry for section 13, restore after.
  3. Section 14's "thread index is plaintext" assertion was stale —
     crypto.js SENSITIVE_PATTERNS explicitly lists
     `^labcharts-[^-]+-chat-threads$`. Real profile ids (createProfile →
     Date.now().toString(36), hyphen-free) match it, so the index IS
     encrypted. Corrected the assertion to match crypto.js.

Vitest: 72 → 75 tests, ~4.57s → ~5.0s.
Puppeteer originals: 28 → 25 remaining.
Full run-tests.sh: green.
